# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SH3 Backend API",
        "description": "API de Atendimentos feita por Hugo Galli para SH3",
        "contact": {
            "name": "Hugo Galli",
            "email": "hugo.galli.r@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/clientes": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Lista de clientes",
                "description": "Retorna a lista de clientes",
                "operationId": "de55d50d1b708d9e5c63ee3f8fe8f54a",
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "clientes": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Cliente"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "jwt_token": []
                    }
                ]
            }
        },
        "/api/cliente": {
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Criar um novo cliente",
                "description": "Cria um novo cliente",
                "operationId": "469f4d88a1d27b95b61b0922e033d8f3",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Título do cliente",
                                        "type": "string",
                                        "example": "Novo Cliente"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Cliente criado com sucesso"
                                        },
                                        "cliente": {
                                            "$ref": "#/components/schemas/Cliente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "jwt_token": []
                    }
                ]
            }
        },
        "/api/cliente/{id}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Detalhes de um cliente específico",
                "description": "Retorna os detalhes de um cliente específico",
                "operationId": "813f501e73b90e51802f8a9400987353",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "cliente": {
                                            "$ref": "#/components/schemas/Cliente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Cliente não encontrado"
                    }
                },
                "security": [
                    {
                        "jwt_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Atualizar um cliente existente",
                "description": "Atualiza um cliente existente",
                "operationId": "6be039ccbab018c4c6a738085884af79",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Novo título do cliente",
                                        "type": "string",
                                        "example": "Novo Título"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Cliente atualizado com sucesso"
                                        },
                                        "cliente": {
                                            "$ref": "#/components/schemas/Cliente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Cliente não encontrado"
                    }
                },
                "security": [
                    {
                        "jwt_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Excluir um cliente",
                "description": "Exclui um cliente",
                "operationId": "3a430f92cf9e8f7d799e351d936f8e0e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Cliente excluído com sucesso"
                                        },
                                        "cliente": {
                                            "$ref": "#/components/schemas/Cliente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Cliente não encontrado"
                    }
                },
                "security": [
                    {
                        "jwt_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Cliente": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Cliente A"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "jwt_token": {
                "type": "apiKey",
                "scheme": "bearer",
                "description": "Insira valores no formato (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}